+++ src/uart_utils.h
@@ -11,6 +11,7 @@
 #include "menu.h"
 #include "uart_storage.h"
 #include <stdbool.h> 
+#include <momentum/momentum.h>
 #define TEXT_BOX_STORE_SIZE (4096)  // 4KB text box buffer size
 #define RX_BUF_SIZE 2048
 #define PCAP_BUF_SIZE 4096
@@ -26,6 +27,9 @@
 #define PCAP_GLOBAL_HEADER_SIZE 24
 #define PCAP_PACKET_HEADER_SIZE 16
 #define PCAP_TEMP_BUFFER_SIZE 4096
+#define UART_CH_ESP (momentum_settings.uart_esp_channel)
+#define UART_CH_GPS (momentum_settings.uart_nmea_channel)
+#define BAUDRATE (115200)
 
 
 void update_text_box_view(AppState* state);
@@ -61,6 +65,8 @@ typedef struct {
 
 typedef struct UartContext {
     FuriHalSerialHandle* serial_handle;
+    FuriHalSerialHandle* gps_handle;
+    FuriStreamBuffer* gps_stream;
     FuriThread* rx_thread;
     FuriStreamBuffer* rx_stream;
     FuriStreamBuffer* pcap_stream;